# sogapeint-CRM/sogapaint-crm/src/assets/swagger-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API Sogapeint",
      "version": "1.0.0",
      "description": "API backend de l'application Sogapeint."
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/companies": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/companiesNames": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/company/search": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "q",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/company/{companyId}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "companyId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "put": {
          "description": "",
          "parameters": [
            {
              "name": "companyId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "names": {
                    "example": "any"
                  },
                  "normalized_name": {
                    "example": "any"
                  },
                  "abbreviation": {
                    "example": "any"
                  },
                  "industry": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "postalCode": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "website": {
                    "example": "any"
                  },
                  "employees": {
                    "example": "any"
                  },
                  "documents": {
                    "example": "any"
                  },
                  "contractsAsCustomer": {
                    "example": "any"
                  },
                  "contractsAsContact": {
                    "example": "any"
                  },
                  "contractsAsExternalContributor": {
                    "example": "any"
                  },
                  "additionalFields": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "",
          "parameters": [
            {
              "name": "companyId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/company": {
        "post": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "names": {
                    "example": "any"
                  },
                  "normalized_name": {
                    "example": "any"
                  },
                  "abbreviation": {
                    "example": "any"
                  },
                  "industry": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "postalCode": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "website": {
                    "example": "any"
                  },
                  "employees": {
                    "example": "any"
                  },
                  "documents": {
                    "example": "any"
                  },
                  "contractsAsCustomer": {
                    "example": "any"
                  },
                  "contractsAsContact": {
                    "example": "any"
                  },
                  "contractsAsExternalContributor": {
                    "example": "any"
                  },
                  "additionalFields": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "400": {
              "description": "Bad Request"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "/companiesAbbreviations": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "/companies/companies": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/companies/companiesNames": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/companies/company/search": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "q",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/companies/company/{companyId}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "companyId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "put": {
          "description": "",
          "parameters": [
            {
              "name": "companyId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "names": {
                    "example": "any"
                  },
                  "normalized_name": {
                    "example": "any"
                  },
                  "abbreviation": {
                    "example": "any"
                  },
                  "industry": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "postalCode": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "website": {
                    "example": "any"
                  },
                  "employees": {
                    "example": "any"
                  },
                  "documents": {
                    "example": "any"
                  },
                  "contractsAsCustomer": {
                    "example": "any"
                  },
                  "contractsAsContact": {
                    "example": "any"
                  },
                  "contractsAsExternalContributor": {
                    "example": "any"
                  },
                  "additionalFields": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "",
          "parameters": [
            {
              "name": "companyId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/companies/company": {
        "post": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "names": {
                    "example": "any"
                  },
                  "normalized_name": {
                    "example": "any"
                  },
                  "abbreviation": {
                    "example": "any"
                  },
                  "industry": {
                    "example": "any"
                  },
                  "address": {
                    "example": "any"
                  },
                  "city": {
                    "example": "any"
                  },
                  "postalCode": {
                    "example": "any"
                  },
                  "country": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "website": {
                    "example": "any"
                  },
                  "employees": {
                    "example": "any"
                  },
                  "documents": {
                    "example": "any"
                  },
                  "contractsAsCustomer": {
                    "example": "any"
                  },
                  "contractsAsContact": {
                    "example": "any"
                  },
                  "contractsAsExternalContributor": {
                    "example": "any"
                  },
                  "additionalFields": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "400": {
              "description": "Bad Request"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/companies/companiesAbbreviations": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
    }
  }